rewrite type parameters with their manifests
The old implementation didn't rewrite type parameters with their
manifests, also if the type constructor had a manifest, then the whole
type was actually rewritten with the manifest of type constuctor. This
worked nice with algebraic data types, e.g., if `('a,'b) t = 'a * 'b`
then it is ok if the whole type is rewritten with the manifest. But if
it is a type abbreviation (or other syntactically non-injective type
function), then is wrong, cf., `('a,'b) t = ('a,'b) list`, that will be
rewritten to just `list`, so we loose information about `'a` and `'b`.

Given the following types:
```
        type foo = int * int
        type bar = string list
        type ('a,'b) t = ('a * 'b) list
        type check = (foo,bar) t
```

The old implementation will manifest it to `list`, and the newer to
`int,int,string,list,list`