strip hidden modules
There is a `-hide` option in ocamldoc that requires the generator to
strip specified module names from the fully qualified identifiers. Argot
generally obyes this, except for the searching. One need to specify a
full type, e.g., `Core_kernel.Std.in_channel`, even if it is displayed
in documentation as `in_channel`, when `-hide=Core_kernel.Std` is passed
to the command line.