add type unification for manifest searching
This patch introduces a new feature. It allows to establish type
equalities, during manifest based search. Basically, it pushes the
manifest search even further. In practice it allows to refer to a
type using any of its aliases.

To achieve this we use union-find data structure, to collect the
equivalence relation between types. For each type definition, we
put into the same equivalence class the type, corresponding type
expression, and the type manifest. We're doing this on a purely
syntactical level (i.e., on the string representation of a type),
to make things simple.

Once the equivalence is established. We emit it using `add_manifest`
function, using the representative element, as the type manifest.
Currently, we're using `Set.max_elt` as a representative element,
but it is a question, how to choose the best, and what criteria should
be considered to choose the best. Also, experiments show, that
it still matters, that contradicts to the theory.