added -hide-undocumented-values option
Once chosen, the option will prune undocumented values from the html
output. This is useful to get rid of the autogenerated stuff.

Other syntactic entities (including methods) are preserved.

TODO: values are pruned only from the html, but are still part of the
module, so they can still be found, but the link will not work. Possibly
we should prune them away, even from the module list.